requirements.txt
# Core ML/AI libraries
torch>=2.0.0
numpy>=1.21.0
vllm>=0.4.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Audio processing
librosa>=0.9.0
soundfile>=0.12.0

# Text embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
accelerate>=0.20.0

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.28.0

# Triton backend utilities
# Note: triton_python_backend_utils is provided by Triton server container

# Additional utilities for multimodal processing
regex>=2023.0.0 